Move RootPathResolver to format module
