+++ src/scikit_build_core/builder/get_requires.py
@@ -5,13 +5,13 @@ import functools
 import importlib.util
 import os
 import sysconfig
-from pathlib import Path
 from typing import TYPE_CHECKING, Literal
 
 from packaging.tags import sys_tags
 
 from .._compat import tomllib
 from .._logging import logger
+from ..format import pyproject_format
 from ..program_search import (
     best_program,
     get_cmake_programs,
@@ -65,28 +65,6 @@ def _load_scikit_build_settings(
     config_settings: Mapping[str, list[str] | str] | None = None,
 ) -> ScikitBuildSettings:
     return SettingsReader.from_file("pyproject.toml", config_settings).settings
-
-
-@dataclasses.dataclass()
-class RootPathResolver:
-    """Handle ``{root:uri}`` like formatting similar to ``hatchling``."""
-
-    path: Path = dataclasses.field(default_factory=Path)
-
-    def __post_init__(self) -> None:
-        self.path = self.path.resolve()
-
-    def __format__(self, fmt: str) -> str:
-        command, _, rest = fmt.partition(":")
-        if command == "parent":
-            parent = RootPathResolver(self.path.parent)
-            return parent.__format__(rest)
-        if command == "uri" and rest == "":
-            return self.path.as_uri()
-        if command == "" and rest == "":
-            return str(self)
-        msg = f"Could not handle format: {fmt}"
-        raise ValueError(msg)
 
 
 @dataclasses.dataclass(frozen=True)
@@ -164,7 +142,11 @@ class GetRequires:
             return
 
         for build_require in self.settings.build.requires:
-            yield build_require.format(root=RootPathResolver())
+            yield build_require.format(
+                **pyproject_format(
+                    settings=self.settings,
+                )
+            )
 
         for dynamic_metadata in self.settings.metadata.values():
             if "provider" in dynamic_metadata:
